second

+++ ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: ysakine <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/01 13:23:17 by ysakine           #+#    #+#             */
-/*   Updated: 2021/11/02 14:03:59 by ysakine          ###   ########.fr       */
+/*   Updated: 2021/11/03 16:37:46 by ysakine          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,12 +14,12 @@
 
 void	*ft_memset(void *b, int c, size_t len)
 {
-	size_t			i;
-	unsigned char	*r;
+	long long int		i;
+	unsigned char		*r;
 
 	i = 0;
 	r = (unsigned char *)b;
-	while (i < len)
+	while (i < (long long int)len)
 	{
 		r[i] = (unsigned char) c;
 		i++;
